+++ app/src/main/java/com/w4eret1ckrtb1tch/focusstart/domain/model/Currency.kt
@@ -2,14 +2,16 @@ package com.w4eret1ckrtb1tch.focusstart.domain.model
 
 import androidx.annotation.Keep
 import java.io.Serializable
+import java.math.BigDecimal
+import java.math.BigInteger
 
 @Keep
 data class Currency(
     val charCode: String,
     val id: String,
     val name: String,
-    val nominal: Int,
+    val nominal: BigInteger,
     val numCode: String,
-    val previous: Double,
-    var value: Double
+    val previous: BigDecimal,
+    var value: BigDecimal
 ) : Serializable
