+++ app/src/main/java/com/w4eret1ckrtb1tch/focusstart/presentation/item/ItemViewModel.kt
@@ -8,19 +8,20 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.w4eret1ckrtb1tch.focusstart.domain.model.Currency
 import dagger.hilt.android.lifecycle.HiltViewModel
+import java.math.BigDecimal
 import javax.inject.Inject
 
 @HiltViewModel
 class ItemViewModel @Inject constructor() : ViewModel() {
 
     private val currency: MutableLiveData<Currency> = MutableLiveData()
-    private val deviationRate: MutableLiveData<Double> = MutableLiveData(0.0)
-    private val amountCurrency: MutableLiveData<Double> = MutableLiveData(0.0)
+    private val deviationRate: MutableLiveData<BigDecimal> = MutableLiveData(BigDecimal(0.0))
+    private val amountCurrency: MutableLiveData<BigDecimal> = MutableLiveData(BigDecimal(0.0))
     private var textWatcher: TextWatcher? = null
 
     fun getCurrency(): LiveData<Currency> = currency
-    fun getDeviationRate(): LiveData<Double> = deviationRate
-    fun getAmountCurrency(): LiveData<Double> = amountCurrency
+    fun getDeviationRate(): LiveData<BigDecimal> = deviationRate
+    fun getAmountCurrency(): LiveData<BigDecimal> = amountCurrency
 
     fun setCurrency(valueCurrency: Currency) {
         currency.value = valueCurrency
@@ -31,15 +32,26 @@ class ItemViewModel @Inject constructor() : ViewModel() {
         if (textWatcher == null) {
             Log.d("TAG", "initTextWatcher: ok")
             textWatcher = object : TextWatcher {
-                override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {}
+                override fun beforeTextChanged(
+                    s: CharSequence?,
+                    start: Int,
+                    count: Int,
+                    after: Int
+                ) {
+                }
 
-                override fun onTextChanged(input: CharSequence?, start: Int, before: Int, count: Int) {
+                override fun onTextChanged(
+                    input: CharSequence?,
+                    start: Int,
+                    before: Int,
+                    count: Int
+                ) {
                     input?.let {
                         if (it.isEmpty() || it.matches("^\\.".toRegex())) {
-                            amountCurrency.value = 0.0
+                            amountCurrency.value = BigDecimal(0.0)
                         } else {
                             amountCurrency.value =
-                                (it.toString().toDouble()) / currency.value!!.value
+                                (it.toString().toBigDecimal()) / currency.value!!.value
                         }
                     }
                 }
